findmissing: make gce-startup.sh to be able run multiple times
Make gce-startup.sh to be able run multiple times without error
messages.

BUG=b:240879180
TEST=./findmissing/scripts/gce/gce-startup.sh &&
     ./findmissing/scripts/gce/full_sync_and_patch_create.sh

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3789452
Commit-Queue: Tzung-Bi Shih <tzungbi@chromium.org>
Tested-by: Tzung-Bi Shih <tzungbi@chromium.org>

+++ contrib/findmissing/scripts/gce/gce-startup.sh
@@ -15,6 +15,14 @@ WORKSPACE="${HOME}/findmissing_workspace"
 FINDMISSING="${WORKSPACE}/dev-util/contrib/findmissing"
 DATABASE=us-central1:linux-patches-mysql-8
 
+if [[ -e /etc/systemd/system/cloud-sql-proxy.service ]]; then
+  sudo systemctl stop cloud-sql-proxy
+fi
+
+if [[ -e /etc/systemd/system/git-cookie-authdaemon.service ]]; then
+  sudo systemctl stop git-cookie-authdaemon
+fi
+
 # Install Stackdriver logging agent
 curl -sSO https://dl.google.com/cloudagents/install-logging-agent.sh
 sudo bash install-logging-agent.sh
@@ -29,17 +37,27 @@ sudo apt-get install -yq \
   libssl-dev libffi-dev nginx
 
 # Fetch git-cookie-authdaemon to authenticate gerrit api requests
-git -C "${WORKSPACE}" clone https://gerrit.googlesource.com/gcompute-tools
+if [[ -e ${WORKSPACE}/gcompute-tools ]]; then
+  git -C "${WORKSPACE}/gcompute-tools" pull
+else
+  git -C "${WORKSPACE}" clone https://gerrit.googlesource.com/gcompute-tools
+fi
 
 # Fetch source code
-git -C "${WORKSPACE}" clone https://chromium.googlesource.com/chromiumos/platform/dev-util
-ln -sf "${FINDMISSING}" "${WORKSPACE}/findmissing"
+if [[ -e ${WORKSPACE}/dev-util ]]; then
+  git -C "${WORKSPACE}/dev-util" pull
+else
+  git -C "${WORKSPACE}" clone https://chromium.googlesource.com/chromiumos/platform/dev-util
+  ln -sf "${FINDMISSING}" "${WORKSPACE}/findmissing"
+fi
 
 # cloud_sql_proxy requires a secret file which can be retrieved via gcloud
 # Note: this will generate a token that lasts forever (year 9999)
-gcloud iam service-accounts keys \
-  create "${WORKSPACE}/secrets/linux_patches_robot_key.json" \
-  --iam-account=linux-patches-robot@chromeos-missing-patches.google.com.iam.gserviceaccount.com
+if [[ ! -e ${WORKSPACE}/secrets/linux_patches_robot_key.json ]]; then
+  gcloud iam service-accounts keys \
+    create "${WORKSPACE}/secrets/linux_patches_robot_key.json" \
+    --iam-account=linux-patches-robot@chromeos-missing-patches.google.com.iam.gserviceaccount.com
+fi
 
 sudo wget https://dl.google.com/cloudsql/cloud_sql_proxy.linux.amd64 -O /usr/bin/cloud_sql_proxy
 sudo chmod a+x /usr/bin/cloud_sql_proxy
@@ -61,6 +79,7 @@ sudo sh -c "cat >/etc/logrotate.d/findmissing" <<EOF
 EOF
 
 # Python environment setup
+rm -rf "${FINDMISSING}/env"
 python3 -m venv "${FINDMISSING}/env"
 source "${FINDMISSING}/env/bin/activate"
 "${FINDMISSING}/env/bin/pip" install -r "${FINDMISSING}/requirements.txt"
@@ -98,7 +117,7 @@ After=syslog.target network-online.target
 User=${USER}
 Type=simple
 Environment="HOME=${HOME}"
-ExecStart=${WORKSPACE}/gcompute-tools/git-cookie-authdaemon
+ExecStart=${WORKSPACE}/gcompute-tools/git-cookie-authdaemon --nofork
 Restart=on-failure
 RestartSec=10
 KillMode=process
@@ -110,6 +129,7 @@ sudo chmod 644 /etc/systemd/system/git-cookie-authdaemon.service
 sudo ln -sf /usr/bin/python3 /usr/bin/python
 
 # Start service now
+sudo systemctl daemon-reload
 sudo systemctl start cloud-sql-proxy
 sudo systemctl start git-cookie-authdaemon
